[ADD] add coverage and sanitizer to export CMakeTarget, and move header to an aggregated folder to avoid name conflict
