+++ include/reddish/network/connection.h
