+++ test/connection_test.cpp
@@ -1,4 +1,4 @@
-#include <network/connection.h>
+#include <reddish/network/connection.h>
 
 #include <iostream>
